// apps/api/src/modules/inbox/event/ChatSocketEvent.ts
export enum ChatSocketReceivingEvent{
  CUSTOMER_START_CHAT_WITH_GARAGE= "CUSTOMER_START_CHAT_WITH_GARAGE",
  CUSTOMER_START_CHAT_WITH_ADMIN= "CUSTOMER_START_CHAT_WITH_ADMIN",
  GARAGE_START_CHAT_WITH_ADMIN = "GARAGE_START_CHAT_WITH_ADMIN",
  GET_LIST_CHANNEL = "GET_LIST_CHANNEL",
  LISTEN_INBOX = "LISTEN_INBOX",
  MESSAGE = "MESSAGE",
}